# docs/api/openAPI.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "SSI DIM Wallet Stub API",
    "description" : "SSI DIM Wallet Stub API",
    "termsOfService" : "https://www.eclipse.org/legal/termsofuse.php",
    "contact" : {
      "name" : "Eclipse Tractus-X",
      "url" : "https://projects.eclipse.org/projects/automotive.tractusx",
      "email" : "tractusx-dev@eclipse.org"
    },
    "version" : "0.0.1"
  },
  "servers" : [
    {
      "url" : "http://localhost:8080",
      "description" : "Generated server url"
    }
  ],
  "security" : [
    {
      "Authenticate using VP for BDRS directory API" : []
    },
    {
      "Authenticate using access_token" : []
    }
  ],
  "paths" : {
...
    "/api/v2.0.0/credentials" : {
      "post" : {
        "tags" : [
          "APIs consumed by SSI Issuer component"
        ],
        "summary" : "Create a new credential | Store credential for a holder",
        "description" : "New credential sign with issuer wallet and saved in in-memory db and send VC id in the response. | Store credential will give only static response with id.\n",
        "operationId" : "createCredential",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IssueCredentialRequest"
              },
              "examples" : {
                "Create a new credential with signature" : {
                  "description" : "Create a new credential with signature",
                  "value" : {
                    "application" : "catena-x",
                    "payload" : {
                      "issueWithSignature" : {
                        "content" : {
                          "@context" : [
                            "https://www.w3.org/2018/credentials/v1",
                            "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                            "https://w3id.org/security/suites/jws-2020/v1"
                          ],
                          "id" : "did:web:localhost:BPNL000000000000#a1f8ae36-9919-4ed8-8546-535280acc5bf",
                          "type" : [
                            "VerifiableCredential",
                            "BpnCredential"
                          ],
                          "issuer" : "did:web:localhost:BPNL000000000000",
                          "issuanceDate" : "2023-07-19T09:14:45Z",
                          "expirationDate" : "2023-09-30T18:30:00Z",
                          "credentialSubject" : {
                            "bpn" : "BPNL000000000001",
                            "id" : "did:web:localhost:BPNL000000000001",
                            "type" : "BpnCredential"
                          }
                        },
                        "signature" : {
                          "proofMechanism" : "external",
                          "proofType" : "jwt",
                          "keyName" : null
                        }
                      }
                    }
                  }
                },
                "Create a new credential" : {
                  "description" : "Create a new credential",
                  "value" : {
                    "application" : "catena-x",
                    "payload" : {
                      "issue" : {
                        "@context" : [
                          "https://www.w3.org/2018/credentials/v1",
                          "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                          "https://w3id.org/security/suites/jws-2020/v1"
                        ],
                        "id" : "did:web:localhost:BPNL000000000000#a1f8ae36-9919-4ed8-8546-535280acc5bf",
                        "type" : [
                          "VerifiableCredential",
                          "BpnCredential"
                        ],
                        "issuer" : "did:web:localhost:BPNL000000000000",
                        "issuanceDate" : "2023-07-19T09:14:45Z",
                        "expirationDate" : "2023-09-30T18:30:00Z",
                        "credentialSubject" : {
                          "bpn" : "BPNL000000000001",
                          "id" : "did:web:localhost:BPNL000000000001",
                          "type" : "BpnCredential"
                        }
                      }
                    }
                  }
                },
                "Store credential for a holder" : {
                  "description" : "Store credential for a holder",
                  "value" : {
                    "application" : "catena-x-portal",
                    "payload" : {
                      "derive" : {
                        "verifiableCredential" : "eyJraWQiOiJkaWQ6d2ViOnNvbWUtaXNzdWVyI2tleS0xIiwiYWxnIjoiRVMyNTYifQ.eyJzdWIiOiJkZWdyZWVTdWIiLCJhdWQiOiJkaWQ6d2ViOmJkcnMtY2xpZW50IiwibmJmIjoxNzE4MzQzOTAzLCJpc3MiOiJkaWQ6d2ViOnNvbWUtaXNzdWVyIiwiZXhwIjoxNzE4MzQzOTYzLCJpYXQiOjE3MTgzNDM5MDMsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIiwiaHR0cHM6Ly93M2lkLm9yZy9jYXRlbmF4L2NyZWRlbnRpYWxzL3YxLjAuMCJdLCJpZCI6IjFmMzZhZjU4LTBmYzAtNGIyNC05YjFjLWUzN2Q1OTY2ODA4OSIsInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJNZW1iZXJzaGlwQ3JlZGVudGlhbCJdLCJpc3N1ZXIiOiJkaWQ6d2ViOmNvbS5leGFtcGxlLmlzc3VlciIsImlzc3VhbmNlRGF0ZSI6IjIwMjEtMDYtMTZUMTg6NTY6NTlaIiwiZXhwaXJhdGlvbkRhdGUiOiIyMDk5LTA2LTE2VDE4OjU2OjU5WiIsImNyZWRlbnRpYWxTdWJqZWN0Ijp7ImlkIjoiZGlkOndlYjpiZHJzLWNsaWVudCIsImhvbGRlcklkZW50aWZpZXIiOiJCUE5MMDAwMDAwMDAxIn19LCJqdGkiOiJlZDlhNjhkMS0yZjFkLTQxZjgtYWUwOS1hNDBhMTA2OTUwMTUifQ.tdLmrcQpGH-SGBpRpRmFX4AXQJx99uUhDOwuGtSejWkkQ2N_yNtEsoP93xDuBod_AY7zVqY4P_Ofdz-H4zE6nw"
                      }
                    }
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "*/*" : {
                "examples" : {
                  "No VC Type Found Exception" : {
                    "description" : "No VC Type Found Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Unprocessable Entity: No type found in VC",
                      "status" : 422,
                      "detail" : "NoVCTypeFoundException: No type found in VC",
                      "instance" : "/api/v2.0.0/credentials",
                      "properties" : {
                        "timestamp" : 1743157807618
                      }
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Illegal Argument Exception" : {
                    "description" : "Illegal Argument Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Bad request: Invalid token -\u003E token",
                      "status" : 400,
                      "detail" : "IllegalArgumentException: Invalid token -\u003E token",
                      "instance" : "/api/v2.0.0/credentials",
                      "properties" : {
                        "timestamp" : 1743154190590
                      }
                    }
                  },
                  "Parse Stub Exception" : {
                    "description" : "Parse Stub Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Invalid serialized unsecured/JWS/JWE object: Missing second delimiter",
                      "status" : 400,
                      "detail" : "ParseStubException: Invalid serialized unsecured/JWS/JWE object: Missing second delimiter",
                      "instance" : "/api/v2.0.0/credentials",
                      "properties" : {
                        "timestamp" : 1743156805552
                      }
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Internal Error Exception" : {
                    "description" : "Internal Error Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Internal Server Error",
                      "status" : 500,
                      "detail" : "InternalErrorException: ...",
                      "instance" : "/api/v2.0.0/credentials",
                      "properties" : {
                        "timestamp" : 1743062000750
                      }
                    }
                  }
                }
              }
            }
          },
          "201" : {
            "description" : "JWT presentation",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Create a new credential" : {
                    "description" : "Create a new credential",
                    "value" : {
                      "id" : "1f36af58-0fc0-4b24-9b1c-e37d59668089"
                    }
                  },
                  "Create a new credential with signature" : {
                    "description" : "Create a new credential with signature",
                    "value" : {
                      "id" : "1f36af58-0fc0-4b24-9b1c-e37d59668089",
                      "jwt" : "eyJraWQiOiJkaWQ6d2ViOnNvbWUtaXNzdWVyI2tleS0xIiwiYWxnIjoiRVMyNTYifQ.eyJzdWIiOiJkZWdyZWVTdWIiLCJhdWQiOiJkaWQ6d2ViOmJkcnMtY2xpZW50IiwibmJmIjoxNzE4MzQzOTAzLCJpc3MiOiJkaWQ6d2ViOnNvbWUtaXNzdWVyIiwiZXhwIjoxNzE4MzQzOTYzLCJpYXQiOjE3MTgzNDM5MDMsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIiwiaHR0cHM6Ly93M2lkLm9yZy9jYXRlbmF4L2NyZWRlbnRpYWxzL3YxLjAuMCJdLCJpZCI6IjFmMzZhZjU4LTBmYzAtNGIyNC05YjFjLWUzN2Q1OTY2ODA4OSIsInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJNZW1iZXJzaGlwQ3JlZGVudGlhbCJdLCJpc3N1ZXIiOiJkaWQ6d2ViOmNvbS5leGFtcGxlLmlzc3VlciIsImlzc3VhbmNlRGF0ZSI6IjIwMjEtMDYtMTZUMTg6NTY6NTlaIiwiZXhwaXJhdGlvbkRhdGUiOiIyMDk5LTA2LTE2VDE4OjU2OjU5WiIsImNyZWRlbnRpYWxTdWJqZWN0Ijp7ImlkIjoiZGlkOndlYjpiZHJzLWNsaWVudCIsImhvbGRlcklkZW50aWZpZXIiOiJCUE5MMDAwMDAwMDAxIn19LCJqdGkiOiJlZDlhNjhkMS0yZjFkLTQxZjgtYWUwOS1hNDBhMTA2OTUwMTUifQ.tdLmrcQpGH-SGBpRpRmFX4AXQJx99uUhDOwuGtSejWkkQ2N_yNtEsoP93xDuBod_AY7zVqY4P_Ofdz-H4zE6nw"
                    }
                  },
                  "Store credential for a holder" : {
                    "description" : "Store credential for a holder",
                    "value" : {
                      "id" : "1f36af58-0fc0-4b24-9b1c-e37d59668089"
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Missing Request Header Exception" : {
                    "description" : "Missing Request Header Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Please provide the required header: Authorization",
                      "status" : 401,
                      "detail" : "MissingRequestHeaderException: Required request header 'Authorization' for method parameter type String is not present",
                      "instance" : "/api/v2.0.0/credentials",
                      "properties" : {
                        "timestamp" : 1743085396772
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0.0/credentials/{credentialId}" : {
      "patch" : {
        "tags" : [
          "APIs consumed by SSI Issuer component"
        ],
        "summary" : "Sign a credential / Revoke a credential",
        "description" : "Credential already signed now it will send vc in response. | Revoke a credential will give static response.\n",
        "operationId" : "signOrRevokeCredential",
        "parameters" : [
          {
            "name" : "credentialId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignCredentialRequest"
              },
              "examples" : {
                "Sign credential" : {
                  "description" : "Sign credential",
                  "value" : {
                    "sign" : {
                      "proofMechanism" : "external",
                      "proofType" : "jwt"
                    }
                  }
                },
                "Revoke credential" : {
                  "description" : "Revoke credential",
                  "value" : {
                    "payload" : {
                      "revoke" : true
                    }
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "JWT presentation",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Sign credential" : {
                    "description" : "Sign credential",
                    "value" : {
                      "jwt" : "eyJraWQiOiJkaWQ6d2ViOnNvbWUtaXNzdWVyI2tleS0xIiwiYWxnIjoiRVMyNTYifQ.eyJzdWIiOiJkZWdyZWVTdWIiLCJhdWQiOiJkaWQ6d2ViOmJkcnMtY2xpZW50IiwibmJmIjoxNzE4MzQzOTAzLCJpc3MiOiJkaWQ6d2ViOnNvbWUtaXNzdWVyIiwiZXhwIjoxNzE4MzQzOTYzLCJpYXQiOjE3MTgzNDM5MDMsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIiwiaHR0cHM6Ly93M2lkLm9yZy9jYXRlbmF4L2NyZWRlbnRpYWxzL3YxLjAuMCJdLCJpZCI6IjFmMzZhZjU4LTBmYzAtNGIyNC05YjFjLWUzN2Q1OTY2ODA4OSIsInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJNZW1iZXJzaGlwQ3JlZGVudGlhbCJdLCJpc3N1ZXIiOiJkaWQ6d2ViOmNvbS5leGFtcGxlLmlzc3VlciIsImlzc3VhbmNlRGF0ZSI6IjIwMjEtMDYtMTZUMTg6NTY6NTlaIiwiZXhwaXJhdGlvbkRhdGUiOiIyMDk5LTA2LTE2VDE4OjU2OjU5WiIsImNyZWRlbnRpYWxTdWJqZWN0Ijp7ImlkIjoiZGlkOndlYjpiZHJzLWNsaWVudCIsImhvbGRlcklkZW50aWZpZXIiOiJCUE5MMDAwMDAwMDAxIn19LCJqdGkiOiJlZDlhNjhkMS0yZjFkLTQxZjgtYWUwOS1hNDBhMTA2OTUwMTUifQ.tdLmrcQpGH-SGBpRpRmFX4AXQJx99uUhDOwuGtSejWkkQ2N_yNtEsoP93xDuBod_AY7zVqY4P_Ofdz-H4zE6nw"
                    }
                  },
                  "Revoke credential " : {
                    "description" : "Revoke credential ",
                    "value" : {
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Internal Error Exception" : {
                    "description" : "Internal Error Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Internal Server Error",
                      "status" : 500,
                      "detail" : "InternalErrorException: ...",
                      "instance" : "/api/v2.0.0/credentials/a",
                      "properties" : {
                        "timestamp" : 1743062000750
                      }
                    }
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Credential Not Found Exception" : {
                    "description" : "Credential Not Found Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Not Found: No credential found for credentialId -\u003E S",
                      "status" : 404,
                      "detail" : "CredentialNotFoundException: No credential found for credentialId -\u003E S",
                      "instance" : "/api/v2.0.0/credentials/a",
                      "properties" : {
                        "timestamp" : 1743154020889
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0.0/credentials/{externalCredentialId}" : {
      "get" : {
        "tags" : [
          "APIs consumed by SSI Issuer component"
        ],
        "summary" : "Get a credential by external ID",
        "description" : "Get a credential by external ID. It will work if vc is present in in-memory db.\n",
        "operationId" : "getCredential",
        "parameters" : [
          {
            "name" : "externalCredentialId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "Get credential",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Get credential" : {
                    "description" : "Get credential",
                    "value" : {
                      "verifiableCredential" : "eyJraWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwI2MzOTMyZmY1LThkYTQtM2RlOS1hOTQyLTYyMTI1ZjM5NGU0MSIsInR5cCI6IkpXVCIsImFsZyI6IkVTMjU2SyJ9.eyJhdWQiOlsiZGlkOndlYjpsb2NhbGhvc3Q6QlBOTDAwMDAwMDAwMDAwMCIsImRpZDp3ZWI6bG9jYWxob3N0OkJQTkwwMDAwMDAwMDAwMDEiXSwiYnBuIjoiQlBOTDAwMDAwMDAwMDAwMSIsInN1YiI6ImRpZDp3ZWI6bG9jYWxob3N0OkJQTkwwMDAwMDAwMDAwMDAiLCJpc3MiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwiZXhwIjoxNzE5ODA5NTQ1LCJpYXQiOjE3MTk4MDkyNDUsInZjIjp7Imlzc3VhbmNlRGF0ZSI6IjIwMjMtMDctMTlUMDk6MTQ6NDVaIiwiY3JlZGVudGlhbFN1YmplY3QiOnsiYnBuIjoiQlBOTDAwMDAwMDAwMDAwMSIsImlkIjoiZGlkOndlYjpsb2NhbGhvc3Q6QlBOTDAwMDAwMDAwMDAwMSIsInR5cGUiOiJCcG5DcmVkZW50aWFsIn0sImlkIjoiZGlkOndlYjpsb2NhbGhvc3Q6QlBOTDAwMDAwMDAwMDAwMCMxOWNiNjU2Mi1iYWM3LTNkYzMtYWFmNi00NjEyZTM0OWEwMTEiLCJ0eXBlIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIiwiQnBuQ3JlZGVudGlhbCJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsImh0dHBzOi8vY2F0ZW5heC1uZy5naXRodWIuaW8vcHJvZHVjdC1jb3JlLXNjaGVtYXMvYnVzaW5lc3NQYXJ0bmVyRGF0YS5qc29uIiwiaHR0cHM6Ly93M2lkLm9yZy9zZWN1cml0eS9zdWl0ZXMvandzLTIwMjAvdjEiXSwiaXNzdWVyIjoiZGlkOndlYjpsb2NhbGhvc3Q6QlBOTDAwMDAwMDAwMDAwMCIsImV4cGlyYXRpb25EYXRlIjoiMjAyMy0wOS0zMFQxODozMDowMFoifSwianRpIjoiYjcyMWY0NjMtMzM3Yi00MzBhLTkzMDktNjZlNzBjMjNkNTZiIn0._mGVXN4ublBx0-r0lG7_2tSGzwIlhjTWtx-ZFcQMmg4Q9pvF-RnbSDZ0vJLfvWv9egVtFSPE9oqbChCLXVg21g",
                      "credential" : {
                        "issuanceDate" : "2023-07-19T09:14:45Z",
                        "credentialSubject" : {
                          "bpn" : "BPNL000000000001",
                          "id" : "did:web:localhost:BPNL000000000001",
                          "type" : "BpnCredential"
                        },
                        "id" : "did:web:localhost:BPNL000000000000#19cb6562-bac7-3dc3-aaf6-4612e349a011",
                        "type" : [
                          "VerifiableCredential",
                          "BpnCredential"
                        ],
                        "@context" : [
                          "https://www.w3.org/2018/credentials/v1",
                          "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                          "https://w3id.org/security/suites/jws-2020/v1"
                        ],
                        "issuer" : "did:web:localhost:BPNL000000000000",
                        "expirationDate" : "2023-09-30T18:30:00Z"
                      },
                      "revocationStatus" : "false",
                      "signing_key_id" : "did:web:localhost:BPNL000000000000#c3932ff5-8da4-3de9-a942-62125f394e41"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Internal Error Exception" : {
                    "description" : "Internal Error Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Internal Server Error",
                      "status" : 500,
                      "detail" : "InternalErrorException: ...",
                      "instance" : "/api/v2.0.0/credentials/a",
                      "properties" : {
                        "timestamp" : 1743062000750
                      }
                    }
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "examples" : {
                  "Credential Not Found Exception" : {
                    "description" : "Credential Not Found Exception",
                    "value" : {
                      "type" : "about:blank",
                      "title" : "Not Found: No credential found for credentialId -\u003E s",
                      "status" : 404,
                      "detail" : "CredentialNotFoundException: No credential found for credentialId -\u003E s",
                      "instance" : "/api/v2.0.0/credentials/a",
                      "properties" : {
                        "timestamp" : 1743154132008
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CredentialPayload" : {
        "type" : "object",
        "properties" : {
          "issue" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "derive" : {
            "$ref" : "#/components/schemas/StoreRequestDerive"
          },
          "issueWithSignature" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        }
      },
      "IssueCredentialRequest" : {
        "type" : "object",
        "properties" : {
          "application" : {
            "type" : "string"
          },
          "valid" : {
            "type" : "boolean"
          },
          "payload" : {
            "$ref" : "#/components/schemas/CredentialPayload"
          }
        }
      },
      "StoreRequestDerive" : {
        "type" : "object",
        "properties" : {
          "verifiableCredential" : {
            "type" : "string"
          }
        }
      },
...
      "Payload" : {
        "type" : "object",
        "properties" : {
          "revoke" : {
            "type" : "boolean"
          }
        }
      },
      "Sign" : {
        "type" : "object",
        "properties" : {
          "proofType" : {
            "type" : "string"
          },
          "proofMechanism" : {
            "type" : "string"
          }
        }
      },
      "SignCredentialRequest" : {
        "type" : "object",
        "properties" : {
          "sign" : {
            "$ref" : "#/components/schemas/Sign"
          },
          "payload" : {
            "$ref" : "#/components/schemas/Payload"
          }
        }
      }
    },
    "securitySchemes" : {
      "Authenticate using access_token" : {
        "type" : "apiKey",
        "description" : "**Bearer (apiKey)**\nJWT Authorization header using the Bearer scheme.\nEnter **Bearer** [space] and then your token in the text input below.\nExample: Bearer access_token\n",
        "name" : "Authorization",
        "in" : "header"
      },
      "Authenticate using VP for BDRS directory API" : {
        "type" : "apiKey",
        "description" : "**Bearer Token**\nVP of membership VC fo access BDRS directory API. This VP must be generated using this application using query credential API\nExample: Bearer 12345abcdef\n",
        "name" : "Authorization",
        "in" : "header"
      }
    }
  }
}